Implemented simple movement, done testing using the animator for states now.
Conclusion:
There is some nice benefits in terms of QoL by using the animator:
- Less code complexity.
- Easier to switch to a different state
- Easier to add a new state, we don't have to add references to the new state in other scripts
By using the Animator to control our states we do lose some control:
- We can choose between LogicUpdate or PhysicsUpdate, but can't have both.
- If we want one player state to have different animation states we can run into issues with state entry and exit.

Overall the animator is simpler to implement and could make it easier to work with states, but since we loose a lot of control I'm hesitant to use it as we could run into issues with that further down the line. Also we've already implemented a state machine pattern, so it would take time to change over to using the animator instead.
We'll keep using the state pattern in code form as we do now.